extract class and func to file
