+++ src/components/Footer.js
@@ -1,0 +1,31 @@
+import React from 'react';
+
+import FilterLink from './FilterLink';
+
+const Footer = () => {
+  return (
+    <p>
+      Show:
+      {' '}
+      <FilterLink
+        filter='SHOW_ALL'
+      >
+        All
+      </FilterLink>
+      {' '}
+      <FilterLink
+        filter='SHOW_ACTIVE'
+      >
+        Active
+      </FilterLink>
+      {' '}
+      <FilterLink
+        filter='SHOW_COMPLETED'
+      >
+        Completed
+      </FilterLink>
+    </p>
+  );
+};
+
+export default Footer;
