--- src/components/TodoApp.js
@@ -1,191 +1,0 @@
-import React, { Component } from 'react';
-import { store, getVisibleTodos } from '../../main';
-
-let nextTodoId = 0;
-
-const Link = ({
-  active,
-  children,
-  onClick
-}) => {
-  if (active) {
-    return <span>{children}</span>
-  }
-
-  return (
-    <a href="#" onClick={e => {
-      e.preventDefault();
-      onClick();
-    }}>
-      {children}
-    </a>
-  );
-};
-
-class FilterLink extends Component {
-
-  componentDidMount() {
-    this.unsubscribe = store.subscribe(() =>
-      this.forceUpdate()
-    );
-  }
-
-  componentWillUnmount () {
-    this.unsubscribe();
-  }
-
-  render() {
-    const props = this.props;
-    const state = store.getState();
-
-    return (
-      <Link
-        active={
-          props.filter === state.visibilityFilter
-        }
-        onClick={() => {
-          store.dispatch({
-            type: 'SET_VISIBILITY_FILTER',
-            filter: props.filter
-          });
-        }}
-      >
-        {props.children}
-      </Link>
-    );
-  }
-}
-
-const Todo = ({
-  onClick,
-  completed,
-  text
-}) => (
-  <li
-    onClick={onClick}
-    style={{
-      textDecoration:
-        completed ?
-          'line-through' :
-          'none'
-    }}
-  >
-    {text}
-  </li>
-);
-
-const TodoList = ({
-  todos,
-  onTodoClick
-}) => (
-  <ul>
-    {todos.map(todo => 
-      <Todo
-        key={todo.id}
-        {...todo}
-        onClick = {() => onTodoClick(todo.id)}
-      />
-    )}
-  </ul>
-);
-
-class VisibleTodoList extends Component {
-
-  componentDidMount() {
-    this.unsubscribe = store.subscribe(() =>
-      this.forceUpdate()
-    );
-  }
-
-  componentWillUnmount () {
-    this.unsubscribe();
-  }
-  render() {
-    const props = this.pros;
-    const state = store.getState();
-
-    return(
-      <TodoList
-        todos={getVisibleTodos(
-          state.todos,
-          state.visibilityFilter
-        )}
-        onTodoClick={(id) => {
-          store.dispatch({
-            type: 'TOGGLE_TODO',
-            id
-          });
-        }}
-      />
-    );
-  }
-}
-
-const AddTodo = ({
-  onAddClick  
-}) => {
-  let input;
-  return(
-    <div>
-      <input ref={node => {
-        input = node;
-      }} />
-      <button onClick={() => {
-        onAddClick(input.value);
-        input.value = '';
-      }}>
-        Add Todo
-      </button>
-    </div>
-  );
-};
-
-const Footer = () => {
-  return (
-    <p>
-      Show:
-      {' '}
-      <FilterLink
-        filter='SHOW_ALL'
-      >
-        All
-      </FilterLink>
-      {' '}
-      <FilterLink
-        filter='SHOW_ACTIVE'
-      >
-        Active
-      </FilterLink>
-      {' '}
-      <FilterLink
-        filter='SHOW_COMPLETED'
-      >
-        Completed
-      </FilterLink>
-    </p>
-  );
-};
-
-const TodoApp = ({
-  todos,
-  visibilityFilter
-}) => (
-  <div>
-    <AddTodo
-      onAddClick={text => {
-        if(text == '') {
-          return;
-        }
-        store.dispatch({
-          'type': 'ADD_TODO',
-          'id'  : nextTodoId++,
-          text
-        });
-      }}
-    />
-    <VisibleTodoList />
-    <Footer />
-  </div>
-);
-
-export default TodoApp;
