+++ src/components/index.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+
+import AddTodo from './AddTodo';
+import VisibleTodoList from './VisibleTodoList';
+import Footer from './Footer';
+
+const TodoApp = () => (
+  <div>
+    <AddTodo />
+    <VisibleTodoList />
+    <Footer />
+  </div>
+);
+
+export default TodoApp;
